# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Fully control your smart home's Lamps using your mobile app"
  version: "1.0.0"
  title: "Smart bulb API"
host: "ibtikar-gl-bulb.appspot.com"
basePath: "/api"
schemes:
- "https"
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /houses:
    post:
      tags:
      - "HOUSES"
      summary: "Add new House"
      description: "This function allows user to create new House.\n"
      operationId: "housesPOST"
      parameters:
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      - in: "body"
        name: "house object"
        description: "house object"
        required: true
        schema:
          $ref: "#/definitions/houseObject"
      responses:
        200:
          description: "house object."
          schema:
            $ref: "#/definitions/House"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "HOUSES"
  /houses/{HOUSE_ID}:
    get:
      tags:
      - "HOUSES"
      summary: "Get house info"
      description: "This returns you all info related to a house by passing house\
        \ to the\nAPI.\n"
      operationId: "housesHOUSE_IDGET"
      parameters:
      - name: "HOUSE_ID"
        in: "path"
        description: "house id to retrieve info about."
        required: true
        type: "number"
        format: "int32"
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      responses:
        200:
          description: "house object."
          schema:
            $ref: "#/definitions/House"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "HOUSES"
    put:
      tags:
      - "HOUSES"
      summary: "Update house info."
      description: "This function allows user to update his/her house info.\n"
      operationId: "housesHOUSE_IDPUT"
      parameters:
      - name: "HOUSE_ID"
        in: "path"
        description: "houseid to delete"
        required: true
        type: "number"
        format: "int32"
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      - in: "body"
        name: "house object"
        description: "house object"
        required: true
        schema:
          $ref: "#/definitions/houseObject"
      responses:
        200:
          description: "house object."
          schema:
            $ref: "#/definitions/House"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "HOUSES"
    delete:
      tags:
      - "HOUSES"
      summary: "Delete user account."
      description: "This function allows user to delete his/her house.\n"
      operationId: "housesHOUSE_IDDELETE"
      parameters:
      - name: "HOUSE_ID"
        in: "path"
        description: "houseid to delete"
        required: true
        type: "number"
        format: "int32"
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      responses:
        200:
          description: "done"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "HOUSES"
  /rooms:
    post:
      tags:
      - "ROOMS"
      summary: "Add new room to a house"
      description: "This function allows user to create new room inside a house.\n"
      operationId: "roomsPOST"
      parameters:
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      - in: "body"
        name: "room object"
        description: "room object"
        required: true
        schema:
          $ref: "#/definitions/roomObject"
      responses:
        200:
          description: "room object."
          schema:
            $ref: "#/definitions/House"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ROOMS"
  /rooms/{ROOM_ID}:
    get:
      tags:
      - "ROOMS"
      summary: "Get room info"
      description: "This returns you all info related to a room by passing room to\
        \ the\nAPI.\n"
      operationId: "roomsROOM_IDGET"
      parameters:
      - name: "ROOM_ID"
        in: "path"
        description: "room id to retrieve info about."
        required: true
        type: "number"
        format: "int32"
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      responses:
        200:
          description: "room object."
          schema:
            $ref: "#/definitions/Room"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ROOMS"
    put:
      tags:
      - "ROOMS"
      summary: "Update room info."
      description: "This function allows user to update his/her room info.\n"
      operationId: "roomsROOM_IDPUT"
      parameters:
      - name: "ROOM_ID"
        in: "path"
        description: "roomid to update"
        required: true
        type: "number"
        format: "int32"
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      - in: "body"
        name: "room object"
        description: "room object"
        required: true
        schema:
          $ref: "#/definitions/roomObject"
      responses:
        200:
          description: "room object."
          schema:
            $ref: "#/definitions/Room"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ROOMS"
    delete:
      tags:
      - "ROOMS"
      summary: "Delete user account."
      description: "This function allows user to delete his/her room from a house.\n"
      operationId: "roomsROOM_IDDELETE"
      parameters:
      - name: "ROOM_ID"
        in: "path"
        description: "rromid to delete"
        required: true
        type: "number"
        format: "int32"
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      responses:
        200:
          description: "done"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ROOMS"
...
definitions:
...
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
...
  House:
    type: "object"
    properties:
      id:
        type: "number"
        description: "house id"
      name:
        type: "string"
        description: "house name, defined by user."
      secret_ky:
        type: "string"
        description: "secret hash that used to control the whole house devices."
      created_at:
        type: "string"
        description: "DateString that indicates when this house created"
      modified_at:
        type: "string"
        description: "DateString that indicates when this house modified"
  houseObject:
    type: "object"
    required:
    - "hn"
    - "sk"
    - "uid"
    properties:
      hn:
        type: "string"
        description: "house name"
      uid:
        type: "number"
        description: "user id."
      sk:
        type: "string"
        description: "secret key"
    description: "house object"
  Room:
    type: "object"
    properties:
      id:
        type: "number"
        description: "room id"
      name:
        type: "string"
        description: "room name"
      apartment_id:
        type: "number"
        description: "house that room is involved in."
      secret_ky:
        type: "string"
        description: "secret hash that used to control the whole house devices."
      created_at:
        type: "string"
        description: "DateString that indicates when this room created"
      modified_at:
        type: "string"
        description: "DateString that indicates when this room modified"
  roomObject:
    type: "object"
    required:
    - "hid"
    - "rn"
    - "sk"
    - "uid"
    properties:
      hid:
        type: "number"
        description: "house id"
      sk:
        type: "string"
        description: "secret key"
      uid:
        type: "number"
        description: "user id"
      rn:
        type: "string"
        description: "room name"
...